Box IOU matcher - pass int64 pointer to copy data

+++ rocAL/source/meta_data/bounding_box_graph.cpp
@@ -218,14 +218,14 @@ void BoundingBoxGraph::update_box_encoder_meta_data(std::vector<float> *anchors,
     }
 }
 
-void BoundingBoxGraph::update_box_iou_matcher(BoxIouMatcherInfo &iou_matcher_info, int *matches_idx_buffer, pMetaDataBatch full_batch_meta_data) {
+void BoundingBoxGraph::update_box_iou_matcher(BoxIouMatcherInfo &iou_matcher_info, int64_t *matches_idx_buffer, pMetaDataBatch full_batch_meta_data) {
     auto bb_coords_batch = full_batch_meta_data->get_bb_cords_batch();
     unsigned anchors_size = iou_matcher_info.anchors->size() / 4;  // divide the anchors_size by 4 to get the total number of anchors
     BoundingBoxCord *bbox_anchors = reinterpret_cast<BoundingBoxCord *>(iou_matcher_info.anchors->data());
 
-    std::vector<int *> matches(full_batch_meta_data->size());
+    std::vector<int64_t *> matches(full_batch_meta_data->size());
     for (int i = 0; i < full_batch_meta_data->size(); i++) {
-        matches[i] = reinterpret_cast<int *>(matches_idx_buffer + i * anchors_size);
+        matches[i] = reinterpret_cast<int64_t *>(matches_idx_buffer + i * anchors_size);
     }
 
 #pragma omp parallel for
@@ -249,7 +249,7 @@ void BoundingBoxGraph::update_box_iou_matcher(BoxIouMatcherInfo &iou_matcher_inf
                 // Find col maximum in (bb_count x anchors_size) IoU values calculated
                 if (iou_val > matched_vals[anchor_idx]) {
                     matched_vals[anchor_idx] = iou_val;
-                    matches[i][anchor_idx] = static_cast<int>(bb_idx);
+                    matches[i][anchor_idx] = static_cast<int64_t>(bb_idx);
                 }
 
                 // Find row maximum in (bb_count x anchors_size) IoU values calculated
